feat: try to move background image from tmp prefix
